refactor(jellyfish-transaction): move vault from loan to its own file
