+++ packages/jellyfish-transaction/src/script/dftx/dftx.ts
@@ -87,23 +87,25 @@ import {
   SetLoanToken,
   CUpdateLoanToken,
   UpdateLoanToken,
+  TakeLoan,
+  CTakeLoan,
+  CPaybackLoan,
+  PaybackLoan
+} from './dftx_loans'
+import {
   CCreateVault,
   CreateVault,
   CUpdateVault,
   UpdateVault,
-  DepositToVault,
   CDepositToVault,
-  WithdrawFromVault,
+  DepositToVault,
   CWithdrawFromVault,
-  CloseVault,
+  WithdrawFromVault,
   CCloseVault,
-  CTakeLoan,
-  TakeLoan,
-  CPaybackLoan,
-  PaybackLoan,
-  PlaceAuctionBid,
-  CPlaceAuctionBid
-} from './dftx_loans'
+  CloseVault,
+  CPlaceAuctionBid,
+  PlaceAuctionBid
+} from './dftx_vault'
 
 /**
  * DeFi Transaction
