+++ packages/jellyfish-transaction/src/script/dftx/index.ts
@@ -15,6 +15,7 @@ export * from './dftx_pool'
 export * from './dftx_price'
 export * from './dftx_token'
 export * from './dftx_loans'
+export * from './dftx_vault'
 export * from './dftx_unmapped'
 export * from './dftx_icxorderbook'
 export * from './dftx'
